Refactor Launcher Client to work in multi-threaded situations
The first pass of the Launcher Client enabled communication to the
Launcher in a single threaded fashion. One Launcher Client would
exist per process. This change refactors the client into two pieces,
the LauncherCli and LauncherMgr.

The manager's job is to receive requests from clients and proxy
those requests to the Launcher itself. It processes requests in a
serial fashion to ensure only one request to the Launcher is processed
at once.

The client connects to the manager and acts just like it did
previously. The only difference being that there can be any number of
clients per process now.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>